+++ CHANGELOG.md
@@ -39,6 +39,8 @@ All notable changes to this project will be documented in this file.
 - Fix a longstanding problem of including empty `categories`, `shortNames` and `additionalPrinterColumns` in the CRDs,
   which could cause problems with GitOps tools (e.g. ArgoCD) reporting a diff in the custom resources.
   See [our internal issue](https://github.com/stackabletech/hdfs-operator/issues/626) and [the fix](https://github.com/kube-rs/kube/pull/2042) for details ([#975]).
+- The operator now watches all resources that it creates and early-exits the reconcile action when the
+  cluster is marked for deletion ([#985]).
 
 [#961]: https://github.com/stackabletech/nifi-operator/pull/961
 [#966]: https://github.com/stackabletech/nifi-operator/pull/966
@@ -48,6 +50,7 @@ All notable changes to this project will be documented in this file.
 [#976]: https://github.com/stackabletech/nifi-operator/pull/976
 [#982]: https://github.com/stackabletech/nifi-operator/pull/982
 [#984]: https://github.com/stackabletech/nifi-operator/pull/984
+[#985]: https://github.com/stackabletech/nifi-operator/pull/985
 
 ## [26.7.0] - 2026-07-21